+++ src/tests/test_card/test_treasure_seeking_seelie.py
@@ -2,7 +2,7 @@ import unittest
 
 from .common_imports import *
 
-class TestDunyarzad(unittest.TestCase):
+class TestTreasureSeekingSeelie(unittest.TestCase):
     def test_behaviour(self):
         base_state = ACTION_TEMPLATE
         base_state = replace_hand_cards(base_state, Pid.P1, Cards({